getTestTemplate fix
Modified the getTestTemplate function. The check for `test` availability wasn't accurate enough. `var || 'default'` does not cover a lot of episodes. Infact I was getting an `Expected () to start arrow function, but got ';' instead of '=>' ` warning because `test` was a series of blank spaces so it wasn't undefined so it wasn't been converted in `true` and the trim would transform it in an empty string. Lately `return ();` threw the warning.